// src/app/models/employee/employee-get-details.model.ts
export interface EmployeeGetDetailsModel {

  employeeID: number;

  employeeName: string;

  email: string;

  phone: string;

  alternatePhone: string;

  city: string;

  state: string;

  pinCode: string;

  address: string;

  departmentName: string;

  designationID: number;

  designationName: string;

  roleID: number;

  roleName: string;

  isActive: boolean;
}